fix bug: coefficient type

+++ src/shader/config.h
@@ -12,9 +12,7 @@ namespace rayjoin {
 using index_t = uint32_t;
 using polygon_id_t = int;
 using coord_t = double;
-// using coefficient_t = __int128;
-// using coefficient_t = int64_t;
-using coefficient_t = double;
+using coefficient_t = __int128;
 }  // namespace rayjoin
 
 #endif  // RAYJOIN_CONFIG_H
